exploring supplemental data from Ventilation during COVID-19 in a school for students with intellectual and developmental disabilities (IDD)
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0291840

punchline is critical Figure 8b appears to be unreproducible from the published data.
And analyzing the published data appropriately shows no effect. 
The study design is flawed because there should be no effect: CO2 measurements are over a few days in November but COVID over a full year. These should both strongly covary on the timescale of a single day. The averaging of COVID counts and only using a few CO2 measurements at one time should be uncorrlated because the covariance is completely unobservable with this design. It was weird they thought there would be an effect, let alone found one.

I'm sure the mistakes are honest, but this needs to be retracted. Unless the authors can explain why I'm completely misinterpreting their study and how we should correctly interpret it. This is a very important topic and needs to be done right.